--- src/app/context/SesionContext/hooks/index.js
@@ -1,4 +1,0 @@
-import useSession from './useSession'
-import useSessionDispatch from './useSessionDispatch'
-
-export { useSessionDispatch, useSession }
