--- src/app/context/SesionContext/hooks/useSession.js
@@ -1,5 +1,0 @@
-import { useContext } from 'react'
-import { AuthContext } from '../../../services/Auth'
-const useSession = () => useContext(AuthContext)
-
-export default useSession
