+++ src/app/context/SesionContext/index.js
@@ -1,0 +1,2 @@
+import { AuthContext } from './useSession'
+export default AuthContext
