+++ src/app/services/Auth/index.js
@@ -1,5 +1,4 @@
-import { AuthContext } from './Authentification'
 import Login from './Login'
 import SignUp from './SignUp'
 
-export { AuthContext, Login, SignUp }
+export { Login, SignUp }
